moved utils into src
